+++ api/src/resolvers/User.ts
@@ -1,0 +1,13 @@
+import { objectType } from "nexus";
+
+export const User = objectType({
+  name: "User",
+  definition(t) {
+    t.model.id();
+    t.model.email();
+    t.model.name();
+    t.model.picture();
+    t.model.createdAt();
+    t.model.updatedAt();
+  }
+});
