+++ api/src/resolvers/index.ts
@@ -1,0 +1,6 @@
+export * from "./Query";
+export * from "./Mutation";
+
+export * from "./Course";
+export * from "./CourseMessage";
+export * from "./User";
